refactor: adjust tests and comments
This cleans up some test layouts and comments.
