+++ bearer_authorization_challenge.go
@@ -18,17 +18,15 @@ import (
 // for an endpoint acting as an OAuth 2.0 protected resource, and returns the error with its status corrected for that
 // context. The caller writes the response body.
 //
-// r may be nil. The response writers that call this - WriteIntrospectionError and writeClientRegistrationError - are
-// public API taking only a context, and the request is not reliably on it: the handlers store it under
-// RequestContextKey on their own derived context, which a caller writing the error from the context it started with
-// will not have. A nil request means the scheme used cannot be established, which RFC 9449 Section 7.2 covers
-// explicitly - "otherwise, both Bearer and DPoP challenges MAY be used to deliver error information" (Figure 19). The
-// challenge stays conformant either way; supplying the request only buys the more precise placement of Figures 16
-// and 18.
+// r may be nil. The response writers calling this take only a context, and the request is not reliably on it: the
+// handlers store it under RequestContextKey on their own derived context. A nil request means the scheme used cannot
+// be established, which RFC 9449 Section 7.2 Figure 19 covers: "otherwise, both Bearer and DPoP challenges MAY be
+// used to deliver error information". The challenge stays conformant either way; supplying the request only buys the
+// more precise placement of Figures 16 and 18.
 //
-// Status correction: ErrInvalidDPoPProof and ErrUseDPoPNonce are HTTP 400, which is correct at the token endpoint per
-// RFC 9449 Section 5, but Section 7.1 Figure 16 and Section 9 Figure 24 both require 401 at a protected resource.
-// They are promoted here rather than by forking the error variables, which would change public API.
+// ErrInvalidDPoPProof and ErrUseDPoPNonce are HTTP 400, which is correct at the token endpoint per RFC 9449
+// Section 5, but Section 7.1 Figure 16 and Section 9 Figure 24 both require 401 at a protected resource. They are
+// promoted here rather than by forking the error variables, which would change public API.
 func (f *Fosite) WriteBearerAuthorizationChallenge(ctx context.Context, rw http.ResponseWriter, r *http.Request, err error) (rfc *RFC6749Error) {
 	rfc = ErrorToRFC6749Error(err)
 
@@ -59,7 +57,7 @@ func (f *Fosite) WriteBearerAuthorizationChallenge(ctx context.Context, rw http.
 //     client bootstrapping a nonce needs the code to know what to retry.
 //   - When the scheme used can be established unambiguously, only the corresponding challenge carries the error
 //     information (Figure 18 for Bearer, Figure 16 for DPoP).
-//   - When it cannot - r is nil, see WriteBearerAuthorizationChallenge - both carry it (Figure 19).
+//   - When it cannot, because r is nil, both carry it (Figure 19).
 //   - The 'algs' parameter is always offered on the DPoP challenge (SHOULD, Section 7.1).
 //   - A 'scope' parameter is offered on an insufficient_scope challenge (MAY, RFC 6750 Section 3.1).
 func (f *Fosite) bearerAuthorizationChallenge(ctx context.Context, r *http.Request, rfc *RFC6749Error) (value string) {
@@ -137,14 +135,13 @@ func challengeScheme(scheme string, params ...string) (value string) {
 // and so warrants a RFC 6750 Section 3 'WWW-Authenticate' challenge naming the Bearer and DPoP schemes. It is the
 // predicate both protected-resource error writers use to decide whether to emit one.
 //
-// ErrRequestUnauthorized is deliberately excluded even though it is also an authorization failure. It is what a
-// client authentication branch returns, and a caller that authenticated with client credentials was not using either
-// of those schemes - RFC 6749 Section 5.2 calls for a challenge matching the scheme actually attempted, so answering
-// a failed Basic authentication with 'WWW-Authenticate: Bearer' would advertise the wrong one.
+// ErrRequestUnauthorized is excluded even though it is also an authorization failure: it is what a client
+// authentication branch returns, and RFC 6749 Section 5.2 calls for a challenge matching the scheme actually
+// attempted, so answering a failed Basic authentication with 'WWW-Authenticate: Bearer' would advertise the wrong
+// one.
 //
-// ErrInvalidRequest is excluded for a related reason: it is mostly a malformed method or body, which says nothing
-// about how the caller authenticated. That includes the RFC 9449 Section 7.2 multiple-Authorization-header case,
-// where a challenge would add nothing the 400 does not already convey.
+// ErrInvalidRequest is excluded because it is mostly a malformed method or body, which says nothing about how the
+// caller authenticated.
 func IsBearerCredentialError(err error) (is bool) {
 	return errors.Is(err, ErrInvalidToken) ||
 		errors.Is(err, ErrInsufficientScope) ||
