+++ config.go
@@ -424,16 +424,14 @@ type RFC8628DeviceAuthorizeConfigProvider interface {
 }
 
 // RFC7591ClientRegistrationTokenSecretProvider returns the provider for configuring the secret client registration
-// tokens are signed and verified with. It is declared once here, rather than inline wherever a configurator needs
-// it, because it is embedded by more than one interface across packages (RFC7591ClientRegistrationConfigProvider
-// here; HMACCoreStrategyConfigurator in handler/oauth2 and HMACSHAStrategyConfigurator in compose both need it to
-// build a hmac.HMACStrategy for client registration tokens) and three independently-maintained copies of the same
-// two methods and doc comments would drift.
+// tokens are signed and verified with. It is embedded by RFC7591ClientRegistrationConfigProvider here, and by
+// HMACCoreStrategyConfigurator in handler/oauth2 and HMACSHAStrategyConfigurator in compose, both of which need it to
+// build a hmac.HMACStrategy for client registration tokens.
 type RFC7591ClientRegistrationTokenSecretProvider interface {
 	// GetRFC7591ClientRegistrationGlobalSecret returns the secret used to sign client registration tokens. It is
-	// deliberately separate from the global secret: a client management token never expires and RFC 7592 provides no
-	// way to re-issue one, so signing it with the global secret would mean routine rotation of that secret
-	// permanently locked every registered client out of its own registration.
+	// separate from the global secret because a client management token never expires and RFC 7592 provides no way to
+	// re-issue one, so signing it with the global secret would mean routine rotation of that secret permanently
+	// locked every registered client out of its own registration.
 	GetRFC7591ClientRegistrationGlobalSecret(ctx context.Context) (secret []byte, err error)
 
 	// GetRFC7591ClientRegistrationRotatedGlobalSecrets returns the rotated client registration token secrets, which
@@ -480,7 +478,7 @@ type RFC7591ClientRegistrationConfigProvider interface {
 	// client registration endpoint, defaulting to consts.ScopeClientRegistration. The token must carry at least one
 	// of them. They are a second authorization dimension alongside the audience: the audience says a token is for
 	// this endpoint, the scope says its holder may register clients. None of these scopes is ever grantable to a
-	// registered client - see CheckGrantableScopes.
+	// registered client; see CheckGrantableScopes.
 	GetRFC7591ClientRegistrationScopes(ctx context.Context) (scopes []string)
 }
 
@@ -605,16 +603,13 @@ type IntrospectionEndpointClientAuthStrategyProvider interface {
 // introspection endpoint.
 type IntrospectionEndpointClientAuthDisabledProvider interface {
 	// GetIntrospectionEndpointClientAuthDisabled returns true when the introspection endpoint must not accept client
-	// authentication, leaving an Access Token presented as a bearer credential the only way to authorize a call.
+	// authentication, leaving an Access Token presented as a bearer credential the only way to authorize a call. It
+	// is off by default, because turning it on rejects every caller using client authentication.
 	//
 	// RFC 7662 Section 2.1 requires the endpoint to "require some form of authorization", naming client
-	// authentication and a bearer token as the two examples, and leaves the choice open. This narrows it to the
-	// second. It is off by default, because turning it on rejects every caller using the first.
-	//
-	// A deployment turns this on to make one property hold: that authorizing a call to this endpoint always means
-	// presenting a credential carrying a grant, which can be scoped, audienced, bound to a key or a certificate, and
-	// revoked on its own. Client credentials carry none of those - a client registered for introspection can call it
-	// for as long as it holds its secret - so while both methods are available the weaker one sets the bar.
+	// authentication and a bearer token as the two examples and leaving the choice open. Turning this on narrows it
+	// to the second, so authorizing a call always means presenting a credential carrying a grant which can be scoped,
+	// audienced, bound to a key or a certificate, and revoked on its own. Client credentials carry none of those.
 	GetIntrospectionEndpointClientAuthDisabled(ctx context.Context) (disabled bool)
 }
 
@@ -675,8 +670,8 @@ type MTLSConfigProvider interface {
 // ConfirmationConfigProvider is the configuration ApplyConfirmation consults to decide whether a given RFC 7800
 // confirmation method may be asserted in a token or an introspection response.
 //
-// It deliberately restates the two 'enabled' getters rather than embedding DPoPConfigProvider and MTLSConfigProvider,
-// which together carry ten methods where two are needed. Go interfaces are structural, so any configuration already
+// It restates the two 'enabled' getters rather than embedding DPoPConfigProvider and MTLSConfigProvider, which
+// together carry ten methods where two are needed. Go interfaces are structural, so any configuration already
 // satisfying those providers satisfies this one without change.
 type ConfirmationConfigProvider interface {
 	// GetDPoPEnabled returns true if DPoP handling is enabled.
