+++ dpop.go
@@ -57,8 +57,8 @@ func RequestURL(r *http.Request) string {
 		}
 	}
 
-	// A request served by net/http always carries a URL, but this is reached from authorization decisions - the DPoP
-	// 'htu' comparison and the bearer credential audience fallback - where a panic would be a far worse failure mode
+	// A request served by net/http always carries a URL, but this is reached from authorization decisions such as the
+	// DPoP 'htu' comparison and the bearer credential audience fallback, where a panic would be a worse failure mode
 	// than a mismatch. A hand-constructed request with no URL therefore reconstructs to just the scheme and host,
 	// which no legitimately issued audience or proof will match.
 	if r.URL == nil {
@@ -111,10 +111,10 @@ type DPoPStrategy interface {
 	ValidateDPoPNonce(ctx context.Context, nonce string) (err error)
 }
 
-// DPoPResourceStrategy is the resource-server half of the RFC 9449 strategy, which an endpoint accepting an access
-// token as a credential needs and DPoPStrategy - the type GetDPoPStrategy returns - does not declare.
-// *rfc9449.DefaultStrategy implements it; the assertion is made at the point of use rather than by widening
-// DPoPStrategy so a deployment supplying its own strategy is not broken by a method it has no bound tokens to serve.
+// DPoPResourceStrategy is the resource-server half of the RFC 9449 strategy, needed by an endpoint accepting an
+// access token as a credential and not declared by DPoPStrategy, the type GetDPoPStrategy returns.
+// *rfc9449.DefaultStrategy implements it. The assertion is made at the point of use rather than by widening
+// DPoPStrategy, so a deployment supplying its own strategy is not broken by a method it has no bound tokens to serve.
 type DPoPResourceStrategy interface {
 	// ValidateResourceAccess performs the RFC 9449 7.1/7.2 resource-server checks for a DPoP-bound access token. It
 	// verifies the token was presented under the DPoP scheme, that the proof covers this request and this token via
