+++ handler/oauth2/strategy_hmac.go
@@ -23,9 +23,9 @@ import (
 // This is the required way to construct a HMACCoreStrategy, not merely a convenience: it is the only place that
 // initialises the unexported strategy client registration tokens are signed and verified with, kept on its own
 // secret so that rotating the global secret never invalidates a client management token, which cannot be re-issued.
-// A struct literal compiles - the other collaborators, Enigma and Config, are exported - but leaves that strategy
-// nil, and GenerateClientRegistrationToken/ValidateClientRegistrationToken then return an error naming this
-// constructor rather than panicking or silently falling back to Enigma's global secret.
+// A struct literal compiles, since Enigma and Config are exported, but leaves that strategy nil, and
+// GenerateClientRegistrationToken and ValidateClientRegistrationToken then return an error naming this constructor
+// rather than falling back to Enigma's global secret.
 func NewHMACCoreStrategy(config HMACCoreStrategyConfigurator, prefix string) (strategy *HMACCoreStrategy) {
 	if len(prefix) == 0 || strings.Count(prefix, "%s") != 1 {
 		return &HMACCoreStrategy{
@@ -48,10 +48,8 @@ func NewHMACCoreStrategy(config HMACCoreStrategyConfigurator, prefix string) (st
 // clientRegistrationSecretConfig adapts an HMACCoreStrategyConfigurator so the embedded hmac.HMACStrategy signs and
 // verifies with the client registration token secrets instead of the global ones. Only the two secret getters are
 // overridden, by shadowing the promoted methods of the same name; entropy and hashing continue to come from the
-// embedded configuration unchanged. HMACCoreStrategyConfigurator is embedded directly, rather than alongside a
-// second field of that same type, because it is already a superset of hmac.HMACStrategyConfigurator (the interface
-// hmac.HMACStrategy.Config requires) - a second field would only duplicate the one embedded value and create two
-// places that must always agree.
+// embedded configuration unchanged. HMACCoreStrategyConfigurator is embedded directly because it is already a
+// superset of hmac.HMACStrategyConfigurator, the interface hmac.HMACStrategy.Config requires.
 type clientRegistrationSecretConfig struct {
 	HMACCoreStrategyConfigurator
 }
@@ -65,10 +63,9 @@ func (c *clientRegistrationSecretConfig) GetRotatedGlobalSecrets(ctx context.Con
 }
 
 // errHMACCoreStrategyMissingClientRegistrationSecret is returned by GenerateClientRegistrationToken and
-// ValidateClientRegistrationToken when enigmaClientRegistration is nil - i.e. this HMACCoreStrategy was assembled
-// with a struct literal instead of NewHMACCoreStrategy, so the client registration secret strategy was never wired
-// up. This deliberately does not fall back to Enigma's global secret: that fallback is exactly what enigmaClientRegistration
-// exists to remove.
+// ValidateClientRegistrationToken when enigmaClientRegistration is nil, meaning this HMACCoreStrategy was assembled
+// with a struct literal instead of NewHMACCoreStrategy. It does not fall back to Enigma's global secret, which is the
+// fallback enigmaClientRegistration exists to remove.
 func errHMACCoreStrategyMissingClientRegistrationSecret() error {
 	return errorsx.WithStack(fmt.Errorf("oauth2: HMACCoreStrategy has no client registration token secret strategy configured; construct it with NewHMACCoreStrategy, not a struct literal, so the client registration secret is wired up"))
 }
@@ -150,11 +147,9 @@ func (s *HMACCoreStrategy) IsOpaqueClientRegistrationToken(ctx context.Context, 
 // empty string when the token is not one. An empty signature can never be a legitimate storage key, so a caller may
 // reject on it without a storage round trip.
 //
-// A nil enigmaClientRegistration - this HMACCoreStrategy was assembled with a struct literal instead of
-// NewHMACCoreStrategy - is one such case, and is answered here rather than one layer down: no client registration
-// token can have been minted without that strategy, so no signature it computed could resolve to anything, and the
-// same condition already turns GenerateClientRegistrationToken and ValidateClientRegistrationToken into a
-// diagnosable error.
+// A nil enigmaClientRegistration, meaning this HMACCoreStrategy was assembled with a struct literal instead of
+// NewHMACCoreStrategy, is one such case. No client registration token can have been minted without that strategy, so
+// no signature it computed could resolve to anything.
 func (s *HMACCoreStrategy) ClientRegistrationTokenSignature(ctx context.Context, tokenString string) (signature string) {
 	if s.enigmaClientRegistration == nil {
 		return ""
@@ -183,9 +178,8 @@ func (s *HMACCoreStrategy) GenerateClientRegistrationToken(ctx context.Context, 
 // ValidateClientRegistrationToken validates an RFC 7591 / RFC 7592 client registration token against its session.
 //
 // Unlike ValidateAccessToken there is no fallback to the configured access token lifespan when the session carries no
-// expiry. newClientRegistrationToken always records one - substituting NonExpiringTokenLifespan for a lifespan of zero
-// rather than leaving it unset - so a session reaching here without an expiry is malformed, and treating that as
-// "never expires" is exactly the silent failure that fallback caused for these tokens before.
+// expiry. newClientRegistrationToken always records one, substituting NonExpiringTokenLifespan for a lifespan of
+// zero, so a session reaching here without an expiry is malformed and must not be treated as never expiring.
 func (s *HMACCoreStrategy) ValidateClientRegistrationToken(ctx context.Context, r oauth2.Requester, tokenString string) (err error) {
 	if s.enigmaClientRegistration == nil {
 		return errHMACCoreStrategyMissingClientRegistrationSecret()
