+++ handler/rfc7591/const.go
@@ -10,16 +10,14 @@ import "time"
 const ClientIDEntropy = 64
 
 // NonExpiringTokenLifespan is the lifespan applied to a client registration access token whose configured lifespan is
-// zero, i.e. one documented as never expiring - the normal expectation for an RFC 7592 'registration_access_token',
-// which has no re-issue path.
+// zero, meaning it never expires. That is the normal expectation for an RFC 7592 'registration_access_token', which
+// has no re-issue path.
 //
-// A zero lifespan cannot simply leave the session's 'expires_at' unset: ValidateClientRegistrationToken treats an
-// unset expiry as a sign the session is malformed and rejects it outright, with no fallback to a configured lifespan
-// the way ValidateAccessToken has for ordinary access tokens (see that method's doc comment). Recording an explicit
-// far-future expiry instead is what makes "never expires" mean what it says.
+// A zero lifespan cannot leave the session's 'expires_at' unset, because ValidateClientRegistrationToken treats an
+// unset expiry as malformed and rejects it, with no fallback to a configured lifespan the way ValidateAccessToken has
+// for ordinary access tokens. An explicit far-future expiry is recorded instead.
 //
-// A century is chosen because it is unambiguously beyond any deployment's horizon while remaining far away from
-// time.Time's and time.Duration's limits.
+// A century is beyond any deployment's horizon while remaining far from time.Time's and time.Duration's limits.
 const NonExpiringTokenLifespan = 100 * 365 * 24 * time.Hour
 
 // SectorIdentifierMaxBodyBytes bounds the number of bytes read from a 'sector_identifier_uri' response body. The
