+++ handler/rfc7591/storage.go
@@ -14,13 +14,12 @@ import (
 // ClientRegistrationTokenStorage handles storage requests related to RFC 7592 client management tokens.
 //
 // Client management tokens have their own storage namespace rather than sharing the access token namespace, in the
-// same way access tokens and refresh tokens each have their own. That separation is a security control, not
-// bookkeeping: while these tokens lived in access token storage every consumer of an access token could load one, and
-// two of them - the introspection endpoint's bearer authentication and the RFC 8693 subject_token path - accepted one
-// as a general purpose credential. It is also what the client configuration endpoint's authentication now rests on:
-// DefaultEndpointAuthStrategy resolves a management token from here and never from access token storage, and resolves
-// the client creation token the registration endpoint takes from access token storage and never from here, so neither
-// credential can be presented at the other's endpoint.
+// same way access tokens and refresh tokens each have their own. That separation is a security control: while these
+// tokens lived in access token storage every consumer of an access token could load one, and both the introspection
+// endpoint's bearer authentication and the RFC 8693 subject_token path accepted one as a general purpose credential.
+// The client configuration endpoint's authentication now rests on it, since DefaultEndpointAuthStrategy resolves a
+// management token only from here and the registration endpoint's creation token only from access token storage, so
+// neither credential can be presented at the other's endpoint.
 //
 // An implementation MUST NOT satisfy these methods from the same namespace it uses for access tokens. A management
 // token never expires and authorises full control of the client it names, and a session hydrating here is evaluated
@@ -48,7 +47,7 @@ type Storage interface {
 
 	// AccessTokenStorage is required because a client creation token is an ordinary access token: it is obtained
 	// through the token endpoint and resolved here from the same storage as any other. Client management tokens
-	// remain in ClientRegistrationTokenStorage, and the two endpoints each consult exactly one of the two - which is
-	// what keeps a token of one kind from being presented at the other's endpoint.
+	// remain in ClientRegistrationTokenStorage, and each endpoint consults exactly one of the two namespaces, which
+	// keeps a token of one kind from being presented at the other's endpoint.
 	hoauth2.AccessTokenStorage
 }
