+++ mtls_test.go
@@ -101,8 +101,6 @@ func TestClientCertificateFromRequest(t *testing.T) {
 	})
 
 	t.Run("ShouldPreferTheHeaderOverThePeerCertificate", func(t *testing.T) {
-		// The peer certificate on a connection from a TLS terminating proxy belongs to the proxy. Preferring it would
-		// bind the token to the proxy, which every client behind that proxy would then satisfy.
 		actual, err := ClientCertificateFromRequest(newRequest(other, header, encodeTraefikV3(cert)), header)
 
 		require.NoError(t, err)
@@ -111,8 +109,6 @@ func TestClientCertificateFromRequest(t *testing.T) {
 	})
 
 	t.Run("ShouldIgnoreThePeerCertificateWhenAHeaderIsConfiguredButAbsent", func(t *testing.T) {
-		// A request arriving without the header carries no client certificate, even over a connection that has a peer
-		// certificate of its own: that certificate is the proxy's.
 		actual, err := ClientCertificateFromRequest(newRequest(other, "", ""), header)
 
 		require.NoError(t, err)
